Rewrite MCP stdio protocol
