+++ internal/mcp/framing.go
@@ -1,0 +1,69 @@
+package mcp
+
+import (
+	"bufio"
+	"errors"
+	"fmt"
+	"io"
+	"strconv"
+	"strings"
+)
+
+func readContentLengthMessage(r *bufio.Reader) ([]byte, error) {
+	contentLength := -1
+	for {
+		line, err := r.ReadString('\n')
+		if err != nil {
+			if errors.Is(err, io.EOF) && line == "" {
+				return nil, io.EOF
+			}
+			return nil, err
+		}
+
+		trimmed := strings.TrimRight(line, "\r\n")
+		if trimmed == "" {
+			break
+		}
+
+		n, ok, err := parseContentLengthHeader(trimmed)
+		if err != nil {
+			return nil, err
+		}
+		if ok {
+			contentLength = n
+		}
+	}
+
+	if contentLength < 0 {
+		return nil, fmt.Errorf("missing Content-Length header")
+	}
+
+	payload := make([]byte, contentLength)
+	if _, err := io.ReadFull(r, payload); err != nil {
+		return nil, err
+	}
+	return payload, nil
+}
+
+func writeContentLengthMessage(w io.Writer, payload []byte) error {
+	if _, err := fmt.Fprintf(w, "Content-Length: %d\r\n\r\n", len(payload)); err != nil {
+		return err
+	}
+	_, err := w.Write(payload)
+	return err
+}
+
+func parseContentLengthHeader(line string) (int, bool, error) {
+	key, value, ok := strings.Cut(line, ":")
+	if !ok {
+		return 0, false, fmt.Errorf("invalid MCP header line %q", line)
+	}
+	if !strings.EqualFold(strings.TrimSpace(key), "Content-Length") {
+		return 0, false, nil
+	}
+	n, err := strconv.Atoi(strings.TrimSpace(value))
+	if err != nil || n < 0 {
+		return 0, false, fmt.Errorf("invalid Content-Length %q", value)
+	}
+	return n, true, nil
+}
